cierre de sesion, paso Marca y Modelo a clase y adapto todo el codigo a ello, carga de bbdd de Marcas y Modelos, selección dinamica de Marca y Modelo con ajax. agregar vehiculo no funciona

+++ src/main/java/com/drivedoctor/dominio/ServicioVehiculo.java
@@ -6,7 +6,7 @@ import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 public interface ServicioVehiculo {
 
-    void agregarVehiculo(Long usuarioId,Vehiculo vehiculo) throws UsuarioInexistente;
+    void agregarVehiculo(Integer usuarioId,Vehiculo vehiculo) throws UsuarioInexistente;
 
     List<Vehiculo> verVehiculos(HttpServletRequest request) throws UserSinPermiso;
