+++ src/main/java/com/drivedoctor/infraestructura/RepositorioUsuarioImpl.java
@@ -54,7 +54,7 @@ public class RepositorioUsuarioImpl implements RepositorioUsuario {
     }
 
     @Override
-    public Usuario buscarPorId(Long usuarioId) {
+    public Usuario buscarPorId(Integer usuarioId) {
         return (Usuario) sessionFactory.getCurrentSession().createCriteria(Usuario.class)
                 .add(Restrictions.eq("id", usuarioId))
                 .uniqueResult();
@@ -62,7 +62,7 @@ public class RepositorioUsuarioImpl implements RepositorioUsuario {
 
     @Override
     public List<Vehiculo> getMisVehiculos(Usuario usuario) {
-        Long userid = usuario.getId();
+        Integer userid = usuario.getId();
         Session session = sessionFactory.getCurrentSession();
         Criteria criteria = session.createCriteria(Vehiculo.class);
         criteria.add(Restrictions.eq("usuario", usuario));
