+++ src/main/java/com/drivedoctor/infraestructura/ServicioUsuarioImpl.java
@@ -32,7 +32,7 @@ public class ServicioUsuarioImpl implements ServicioUsuario {
     }
 
     @Override
-    public Usuario buscar(Long usuarioId) {
+    public Usuario buscar(Integer usuarioId) {
         return repositorioUsuario.buscarPorId(usuarioId);
     }
